- Worked on animation render functionality - Implemented endian swap functionality in raw volume importer - Fixed some minor bugs regarding scene import and usage - Minor changes/refactoring

+++ Source/VolumeScatterRenderer/Core/CRandomGenerator.h
@@ -81,7 +81,7 @@ namespace vox
         VOX_DEVICE Vector2f sampleDisk()
         {
             float r = sqrtf(sample1D());
-            float t = 2.0f * M_PI * sample1D();
+            float t = 2.0f * (float)M_PI * sample1D();
 
             return Vector2f(r*cosf(t), r*sinf(t));
         }
